requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
dotenv==0.9.9
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
multidict==6.4.3
packaging==25.0
propcache==0.3.1
PyJWT==2.10.1
pymongo==4.12.1
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
twilio==9.5.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
